[refactor] RequestHandler 를 Controller 와 Service 로 분리
