+++ src/main/java/webserver/RequestHeader.java
@@ -1,11 +1,16 @@
 package webserver;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
 public class RequestHeader {
+    private static final Logger logger = LoggerFactory.getLogger(Request.class);
+
     private final Map<String, String> headers = new HashMap<>();
 
     public RequestHeader(BufferedReader br) throws IOException {
@@ -13,6 +18,7 @@ public class RequestHeader {
              !headerLine.equals("");
              headerLine = br.readLine()
         ) {
+            logger.debug("~~~~~~~~ requestHeader: {} ~~~~~~~~", headerLine);
             String[] headerTokens = headerLine.split(": ");
             if (headerTokens.length == 2) {
                 put(headerTokens);
