+++ src/main/java/webserver/Service.java
@@ -5,51 +5,15 @@ import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.*;
-import java.net.Socket;
+import java.io.DataOutputStream;
+import java.io.File;
+import java.io.IOException;
 import java.nio.file.Files;
 
-import static webserver.RequestMethod.GET;
-import static webserver.RequestMethod.POST;
+public class Service {
+    private static final Logger log = LoggerFactory.getLogger(Service.class);
 
-public class RequestHandler extends Thread {
-    private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
-
-    private final Socket connection;
-
-    public RequestHandler(Socket connectionSocket) {
-        this.connection = connectionSocket;
-    }
-
-    @Override
-    public void run() {
-        log.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
-                connection.getPort());
-
-        try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
-            BufferedReader br = new BufferedReader(new InputStreamReader(in, "UTF-8"));
-            String line = br.readLine();
-            if (line == null) {
-                return;
-            }
-            Request request = new Request(line, br);
-            RequestMethod method = request.getMethod();
-            String url = request.getUrl();
-
-            DataOutputStream dos = new DataOutputStream(out);
-            if (url.equals("/user/create") && method.equals(POST)) {
-                handleUserCreate(request, dos);
-            } else if (url.equals("/user/login") && method.equals(POST)) {
-                handleUserLogin(request, dos);
-            } else if (method.equals(GET)) {
-                handleGet(request, dos);
-            }
-        } catch (IOException e) {
-            log.error(e.getMessage());
-        }
-    }
-
-    private void handleGet(Request request, DataOutputStream dos) throws IOException {
+    public void get(Request request, DataOutputStream dos) throws IOException {
         String url = request.getUrl();
         byte[] body = Files.readAllBytes(new File("./webapp" + url).toPath());
         String contentType = request.getRequestHeader()
@@ -63,7 +27,7 @@ public class RequestHandler extends Thread {
         response.writeTo(dos);
     }
 
-    private void handleUserCreate(Request request, DataOutputStream dos) throws IOException {
+    public void createUser(Request request, DataOutputStream dos) throws IOException {
         RequestBody requestBody = request.getRequestBody();
         if (requestBody == null) {
             return;
@@ -82,7 +46,7 @@ public class RequestHandler extends Thread {
         response.writeTo(dos);
     }
 
-    private void handleUserLogin(Request request, DataOutputStream dos) throws IOException {
+    public void login(Request request, DataOutputStream dos) throws IOException {
         RequestBody requestBody = request.getRequestBody();
         if (requestBody == null) {
             return;
